add: kafka messages for drop of notification ; fix: the randomness to end for a profile for flicks

+++ src/controllers/Likes/toggleLike.ts
@@ -17,6 +17,9 @@ const buildLikeQuery = (user: string, id: string, type: 'quest' | 'comment' | 'f
     if (type === 'quest') query.quest = id;
     return query;
 };
+import { sendNotificationKafka } from "../../utils/sendNotificationKafka"; // example import
+import { QUESTS } from "../../models/Quest/quest.model";
+
 export const toggleLike = async (req: Request, res: Response) => {
     const session = await mongoose.startSession();
     session.startTransaction();
@@ -29,25 +32,50 @@ export const toggleLike = async (req: Request, res: Response) => {
         const { id, type }: QueryParams = req.query as any;
         const query = buildLikeQuery(user, id, type);
         // Find Like document (atomic check)
+        let searched: any = null;
+        // Quest Favorite: Upsert logic
+        if (type === "quest") {
+            const searchedArr = await QUESTS.aggregate([
+                { $match: { _id: new mongoose.Types.ObjectId(id) } },
+                {
+                    $project: {
+                        user: 1,
+                        likeCount: 1,
+                        thumbnailURL: { $arrayElemAt: ["$media.thumbnailURL", 0] }
+                    }
+                }
+            ]).session(session).exec();
+
+            if (!searchedArr || searchedArr.length === 0) {
+                await session.abortTransaction();
+                return handleResponse(res, 404, errors.quest_not_found);
+            }
+            searched = searchedArr[0];
+            if (!searched) {
+                await session.abortTransaction();
+                return handleResponse(res, 404, errors.quest_not_found);
+            }
+            const fav = await QUEST_FAV.findOne({ user, quest: id }).session(session);
+            if (!fav) {
+                await searched.updateOne({ $inc: { likeCount: 1 } }, { session });
+                await QUEST_FAV.create([{ user, quest: id }], { session });
+            } else {
+                await searched.updateOne({ $inc: { likeCount: -1 } }, { session });
+                await fav.deleteOne({ session });
+            }
+        }
+
         const existingLike = await LIKE.findOne(query).session(session);
+
         // Flick logic: Use $inc for atomic likeCount update
         if (type === "flick") {
-            const flick = await FLICKS.findById(id).session(session);
-            if (!flick) {
+            searched = await FLICKS.findById(id, "user thumbnailURL likeCount").session(session);
+            if (!searched) {
                 await session.abortTransaction();
                 return handleResponse(res, 404, errors.flick_not_found);
             }
             const inc = existingLike ? -1 : 1;
             await FLICKS.updateOne({ _id: id }, { $inc: { likeCount: inc } }).session(session);
-        }
-        // Quest Favorite: Upsert logic
-        if (type === "quest") {
-            const fav = await QUEST_FAV.findOne({ user, quest: id }).session(session);
-            if (!fav) {
-                await QUEST_FAV.create([{ user, quest: id }], { session });
-            } else {
-                await fav.deleteOne({ session });
-            }
         }
         // Toggle Like document itself
         if (existingLike) {
@@ -55,8 +83,28 @@ export const toggleLike = async (req: Request, res: Response) => {
         } else {
             await LIKE.create([{ ...query }], { session });
         }
+
         await session.commitTransaction();
         session.endSession();
+
+        if (searched.user !== user) {  // Don't send notification to self
+            const kafkaMessage = {
+                userId: user,
+                contentUserId: searched.user,
+                username: res.locals.username,
+                targetId: id,
+                likeCount  : searched.likeCount,
+                photo: searched.thumbnailURL,
+                targetType: type,
+                action: existingLike ? "unliked" : "liked",
+                timestamp: new Date().toISOString()
+            };
+            // Send Kafka notification asynchronously, don't await so it doesn't delay response
+            await sendNotificationKafka('LIKE_TOGGLE', kafkaMessage).catch((err) => {
+                console.error("Kafka notification error in toggleLike:", err);
+            });
+        }
+
         return handleResponse(res, 200, success.toggle_like);
     } catch (error) {
         await session.abortTransaction();
@@ -64,4 +112,4 @@ export const toggleLike = async (req: Request, res: Response) => {
         console.error(error);
         return handleResponse(res, 500, errors.catch_error);
     }
-}
+};
